// src/hooks/types/contracts/experience.ts
export enum experience {
    createEvent = 'createEvent',
    updateTimeOfEvent = 'updateTimeOfEvent',
    eventApproval = 'eventApproval',
    enterVenueAndBurnTickets = 'enterVenueAndBurnTickets',
    buyTicket = 'buyTicket',
    cancelTickets = 'cancelTickets',
    burnTicketsAndEscrow = 'burnTicketsAndEscrow',
    burnTicketsandRefund = 'burnTicketsandRefund',
    setApprovalForAll = 'setApprovalForAll'
}